+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   when exhausting flush buffer (#70).
 - More consistent name generation across verifier and json parsers
   allowing for namespace wrapped parse/verify/print table functions.
+- Fix unhelpful error on precision loss from float/double conversion
+  in schema and JSON parser.
 
 ## [0.5.0]
 - New schema type aliases: int8, uint8, int16, uint16, int32, uint32,